Moved SelectRegionPlot to grabcut.py
